# fix: Reduce performance test execution times to prevent timeouts